bpm/bonita-core/bonita-process-instance/src/test/java/org/bonitasoft/engine/core/document/api/impl/SDocumentDownloadURLProviderImplTest.java
/**
 * Copyright (C) 2020 Bonitasoft S.A.
 * Bonitasoft, 32 rue Gustave Eiffel - 38000 Grenoble
 * This library is free software; you can redistribute it and/or modify it under the terms
 * of the GNU Lesser General Public License as published by the Free Software Foundation
 * version 2.1 of the License.
 * This library is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
 * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Lesser General Public License for more details.
 * You should have received a copy of the GNU Lesser General Public License along with this
 * program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth
 * Floor, Boston, MA 02110-1301, USA.
 **/
package org.bonitasoft.engine.core.document.api.impl;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.*;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;

/**
 * @author Julien Mege
 */
@RunWith(MockitoJUnitRunner.class)
public class SDocumentDownloadURLProviderImplTest {

    private static final String SERVLET_URL = "documentDownload";

    private SDocumentDownloadURLProvider urlProvider = new SDocumentDownloadURLProviderImpl(SERVLET_URL);

    @Test
    public void generateURL_should_return_encoded_url() {
        assertEquals("documentDownload?fileName=%5BnameWithSpecialChar%5D&contentStorageId=docId",
                urlProvider.generateURL("[nameWithSpecialChar]", "docId"));
    }

}
